Fix submit proposal
